GenAI/Agentic AI/Langchain/Learning/requirements.txt
langchain
langchain_community
pypdf
ipykernel
python-dotenv
bs4
arxiv
pymupdf
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
numpy<2
faiss-cpu
langchain_chroma
scikit-learn
streamlit
pydantic
duckduckgo-search
langchain-experimental